Marvel's Demo UI fixes.
